Make sure, rest-api calls are lower case
Right now, the rest-api is just shoehorned into the normal web page
which is configured for canonical urls. This causes a redirection to
a lower-case-only url for api-requests as well. During such a redirect
however the CORS-header gets lost resulting in an error in the browser
(and a failed request). The recommended fix would be to exclude the
api from canonical-url rules and allow requests as is. Until this is
done however, the temporary fix of just lower-casing every request
must suffice.